pyproject.toml
[tool.poetry]
name = "-online-idea"
version = "0.1.0"
description = ""
authors = ["anton <sparco114@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.2"
Django = "^4.2.7"
requests = "^2.31.0"
lxml = "^4.9.3"
djangorestframework = "^3.14.0"
Markdown = "^3.5.1"
django-filter = "^23.3"
django-cors-headers = "^4.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
